# rigweda/backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rigweda API
  version: 0.8.0
  description: Independently implemented HR and payroll platform API.
servers:
  - url: /api/v1
paths:
...
  /payroll/runs:
    get:
      tags: [Payroll]
      security: [{ bearerAuth: [] }]
      summary: List payroll runs
      responses:
        "200": { description: Payroll runs }
    post:
      tags: [Payroll]
      security: [{ bearerAuth: [] }]
      summary: Create a payroll run
      responses:
        "201": { description: Payroll run created }
  /payroll/runs/{runId}/compute:
    post:
      tags: [Payroll]
      security: [{ bearerAuth: [] }]
      summary: Compute a payroll run from compensation, attendance, and approved adjustments
      parameters:
        - { name: runId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Payroll run computed }
  /payroll/compensations:
    get:
      tags: [Payroll]
      security: [{ bearerAuth: [] }]
      summary: List employee compensation records
      responses:
        "200": { description: Compensation records }
    post:
      tags: [Payroll]
      security: [{ bearerAuth: [] }]
      summary: Create an employee compensation record
      responses:
        "201": { description: Compensation created }
  /payroll/me/payslips:
    get:
      tags: [Payroll]
      security: [{ bearerAuth: [] }]
      summary: List finalized payroll runs available to the authenticated employee
      responses:
        "200": { description: Employee payslips }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...