# rigweda/backend/src/modules/payroll/payroll.routes.js
const express=require("express");
const {asyncHandler}=require("../../middleware/async-handler");
const {validate}=require("../../middleware/validate");
const schemas=require("./payroll.validation");

const createPayrollRouter=(service,permissions)=>{
  const router=express.Router();
  const ok=(req,res,data,status=200)=>res.status(status).json({success:true,data,requestId:req.id});
  router.get("/metadata",validate(schemas.metadataSchema),asyncHandler(async(req,res)=>ok(req,res,await service.metadata(req.auth.organizationId,req.auth.userId,req.validated.query.search))));
  router.get("/me/payslips",permissions.self,validate(schemas.rangeSchema),asyncHandler(async(req,res)=>ok(req,res,await service.myPayslips(req.auth.organizationId,req.auth.userId,req.validated.query))));
  router.get("/me/payslips/:runId/items",permissions.self,validate(schemas.runIdSchema),asyncHandler(async(req,res)=>ok(req,res,await service.myItems(req.auth.organizationId,req.auth.userId,req.validated.params.runId,{page:1,pageSize:100,search:""}))));
  router.get("/me/payslip-items/:itemId",permissions.self,validate(schemas.itemIdSchema),asyncHandler(async(req,res)=>ok(req,res,await service.payslip(req.auth.organizationId,req.auth.userId,req.validated.params.itemId))));
  router.get("/schedules",permissions.manage,validate(schemas.listSetupSchema),asyncHandler(async(req,res)=>ok(req,res,await service.schedules(req.auth.organizationId,req.validated.query))));
  router.post("/schedules",permissions.manage,validate(schemas.createScheduleSchema),asyncHandler(async(req,res)=>ok(req,res,await service.createSchedule(req.auth.organizationId,req.validated.body,req.auth.userId),201)));
  router.patch("/schedules/:scheduleId",permissions.manage,validate(schemas.updateScheduleSchema),asyncHandler(async(req,res)=>ok(req,res,await service.updateSchedule(req.auth.organizationId,req.validated.params.scheduleId,req.validated.body,req.auth.userId))));
  router.get("/components",permissions.manage,validate(schemas.listSetupSchema),asyncHandler(async(req,res)=>ok(req,res,await service.components(req.auth.organizationId,req.validated.query))));
  router.post("/components",permissions.manage,validate(schemas.createComponentSchema),asyncHandler(async(req,res)=>ok(req,res,await service.createComponent(req.auth.organizationId,req.validated.body,req.auth.userId),201)));
  router.patch("/components/:componentId",permissions.manage,validate(schemas.updateComponentSchema),asyncHandler(async(req,res)=>ok(req,res,await service.updateComponent(req.auth.organizationId,req.validated.params.componentId,req.validated.body,req.auth.userId))));
  router.get("/compensations",permissions.manage,validate(schemas.listCompensationSchema),asyncHandler(async(req,res)=>ok(req,res,await service.compensations(req.auth.organizationId,req.validated.query))));
  router.post("/compensations",permissions.manage,validate(schemas.createCompensationSchema),asyncHandler(async(req,res)=>ok(req,res,await service.createCompensation(req.auth.organizationId,req.validated.body,req.auth.userId),201)));
  router.patch("/compensations/:compensationId",permissions.manage,validate(schemas.updateCompensationSchema),asyncHandler(async(req,res)=>ok(req,res,await service.updateCompensation(req.auth.organizationId,req.validated.params.compensationId,req.validated.body,req.auth.userId))));
  router.get("/adjustments",permissions.manage,validate(schemas.rangeSchema),asyncHandler(async(req,res)=>ok(req,res,await service.adjustments(req.auth.organizationId,req.validated.query))));
  router.post("/adjustments",permissions.manage,validate(schemas.createAdjustmentSchema),asyncHandler(async(req,res)=>ok(req,res,await service.createAdjustment(req.auth.organizationId,req.validated.body,req.auth.userId),201)));
  router.post("/adjustments/:adjustmentId/review",permissions.approve,validate(schemas.reviewAdjustmentSchema),asyncHandler(async(req,res)=>ok(req,res,await service.reviewAdjustment(req.auth.organizationId,req.validated.params.adjustmentId,req.validated.body,req.auth.userId))));
  router.get("/runs",validate(schemas.rangeSchema),asyncHandler(async(req,res)=>ok(req,res,await service.runs(req.auth.organizationId,req.validated.query))));
  router.post("/runs",permissions.manage,validate(schemas.createRunSchema),asyncHandler(async(req,res)=>ok(req,res,await service.createRun(req.auth.organizationId,req.validated.body,req.auth.userId),201)));
  router.get("/runs/:runId",validate(schemas.runIdSchema),asyncHandler(async(req,res)=>ok(req,res,await service.run(req.auth.organizationId,req.validated.params.runId))));
  router.get("/runs/:runId/items",validate(schemas.runIdSchema),asyncHandler(async(req,res)=>ok(req,res,await service.items(req.auth.organizationId,req.validated.params.runId,{page:Number(req.query.page)||1,pageSize:Math.min(Number(req.query.pageSize)||50,100),search:String(req.query.search||"")}))));
  router.get("/runs/:runId/export",permissions.export,validate(schemas.runIdSchema),asyncHandler(async(req,res)=>{const csv=await service.export(req.auth.organizationId,req.validated.params.runId);res.set({"Content-Type":"text/csv; charset=utf-8","Content-Disposition":`attachment; filename="payroll-${req.validated.params.runId}.csv"`});res.send(`\uFEFF${csv}`);}));
  router.post("/runs/:runId/compute",permissions.manage,validate(schemas.runIdSchema),asyncHandler(async(req,res)=>ok(req,res,await service.compute(req.auth.organizationId,req.validated.params.runId,req.auth.userId))));
  router.post("/runs/:runId/approve",permissions.approve,validate(schemas.runActionSchema),asyncHandler(async(req,res)=>ok(req,res,await service.approve(req.auth.organizationId,req.validated.params.runId,req.validated.body,req.auth.userId))));
  router.post("/runs/:runId/finalize",permissions.approve,validate(schemas.runActionSchema),asyncHandler(async(req,res)=>ok(req,res,await service.finalize(req.auth.organizationId,req.validated.params.runId,req.validated.body,req.auth.userId))));
  router.post("/runs/:runId/reopen",permissions.approve,validate(schemas.runActionSchema),asyncHandler(async(req,res)=>ok(req,res,await service.reopen(req.auth.organizationId,req.validated.params.runId,req.validated.body,req.auth.userId))));
  router.post("/runs/:runId/void",permissions.manage,validate(schemas.runActionSchema),asyncHandler(async(req,res)=>ok(req,res,await service.void(req.auth.organizationId,req.validated.params.runId,req.validated.body,req.auth.userId))));
  return router;
};
module.exports={createPayrollRouter};